requirements.txt
# base
wheel==0.37.1
# django
Django==4.0.3
# debug toolbar
django-debug-toolbar==3.2.4
# image
Pillow==9.0.1
# crispy forms
django-crispy-forms==1.14.0
# translation
django-modeltranslation==0.17.5
# rich text editor
django-tinymce==3.4.0
# singleton models
django-solo==2.0.0
